Interactive World Map
This Angular application is a world map interface that allows users to interact with different countries on the map. Users can click on countries to retrieve country information pulled from the World Bank API. The interface consists of a map column and an information column, where data is displayed dynamically based on user interactions.